정기데이터 수정 및 추가 모달 적용

+++ app/src/main/java/com/redhorse/accountbank/data/repository/SavePaymentRepository.kt
@@ -40,7 +40,7 @@ class SavePaymentRepository(private val dbHelper: AppDatabaseHelper) {
     }
 
     // 특정 ID의 Payment 가져오기
-    fun getPaymentById(date: String, id: Int): Payment? {
+    fun getPaymentById(date: String, id: Long): Payment? {
         val db = dbHelper.readableDatabase
         val query = "SELECT * FROM $TABLE_SAVE_PAYMENT WHERE id = ?"
         val cursor = db.rawQuery(query, arrayOf(id.toString()))
